third_party/blink/web_tests/platform/generic/fast/dom/Document/document-current-script-expected.txt

Test basic usage of document.currentScript

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<script id="a">
<script id="b">
<script id="c">
<script id="d">
<script id="c">
<script id="e" src="...">
<script id="f">
<script id="f">
<script id="g" src="...">
<script id="h">
<script id="h">
<script id="i">
<script id="j">
<script id="i">
<script id="svg_in_html">
<script id="svg script">
<script id="defer-1" src="..." defer>
<script id="defer-2" src="..." defer>
PASS document.currentScript is null
PASS successfullyParsed is true

TEST COMPLETE
